Home Federal Savings Bank
This assignment was based on the Seinfeld episode where Kramer goes into a bank that is known for giving $100 to any customer if they were not greeted with a "hello." When Kramer is greeted with a "hey," he demands $100. After a debacle of technicalities and "What do we do's", they finally offer him $20 for a greeting that starts with an "H." This code is meant to take in the user's greeting and to reflect that scenario. I learned the startswith() function by doing this assignment.